requirements.txt
opencv-python
pillow
streamlit
numpy
scipy
matplotlib